Fixed warnings. added sanitizers.

+++ metadata/cell.cpp
@@ -1,12 +1,13 @@
 #include "cell.h"
+#include <cstddef>
 
 int TableDefinition::getColumnIdx(std::string columnName)
 {
-    for(int i=0; i<columns.size();i++)
+    for (size_t i = 0; i < columns.size(); i++)
     {
-        if(columns[i].name == columnName)
+        if (columns[i].name == columnName)
         {
-            return i;
+            return (int)i;
         }
     }
     return -1;
@@ -14,9 +15,9 @@ int TableDefinition::getColumnIdx(std::string columnName)
 
 ColumnType TableDefinition::getColumnType(std::string columnName)
 {
-    for(auto& column: columns)
+    for (auto &column : columns)
     {
-        if(column.name == columnName)
+        if (column.name == columnName)
         {
             return column.type;
         }
